[content] Fix retrieving data under 'page' loaded from front matter
