+++ packages/jaspr_content/lib/src/data_loader/data_loader.dart
@@ -6,6 +6,7 @@ import 'dart:convert';
 import 'package:yaml/yaml.dart' as yaml;
 
 import '../page.dart';
+import '../utils.dart';
 
 /// A data loader that loads additional data for a page.
 ///
@@ -28,22 +29,6 @@ abstract class DataLoader {
       return yaml.loadYamlNode(data).normalize();
     } else {
       return data;
-    }
-  }
-}
-
-extension on yaml.YamlNode {
-  Object normalize() {
-    if (this case yaml.YamlMap(:final nodes)) {
-      return {
-        for (final entry in nodes.entries) entry.key.toString(): entry.value.normalize(),
-      };
-    } else if (this case yaml.YamlList(:final nodes)) {
-      return [
-        for (final node in nodes) node.normalize(),
-      ];
-    } else {
-      return value;
     }
   }
 }
